fix MarioSwim regressions

+++ src/Game/Player/MarioSwim.cpp
@@ -1638,7 +1638,7 @@ bool MarioSwim::close() {
     }
 
     setYangleOffset(0.0f);
-    u16 jointID = getAnimator()->getUpperJointID();
+    s32 jointID = getAnimator()->getUpperJointID();
     setJointGlobalMtx(jointID, nullptr);
 
     decideEffect(true);
@@ -2518,7 +2518,7 @@ void MarioSwim::updateTilt() {
     PSMTXRotRad(mtxRotZ, 'z', _E4);
     PSMTXConcat(mtxRotX, mtxRotZ, _B0);
 
-    u16 jointID = getAnimator()->getUpperJointID();
+    s32 jointID = getAnimator()->getUpperJointID();
     setJointGlobalMtx(jointID, _B0);
 }
 
@@ -2839,7 +2839,7 @@ void MarioSwim::decLife() {
         TVec3f gravityVec = -getGravityVec();
         getPlayer()->forceSetHeadVecKeepSide(gravityVec);
 
-        u16 upperJoint = getAnimator()->getUpperJointID();
+        s32 upperJoint = getAnimator()->getUpperJointID();
         setJointGlobalMtx(upperJoint, nullptr);
     }
 }
